requirements.txt
langchain
langchain-core~=0.3.63
langchain-community~=0.3.24
langchain-together
python-dotenv~=1.1.0
beautifulsoup4~=4.13.4
requests~=2.32.3
faiss-cpu
uvicorn~=0.34.2
fastapi~=0.115.12
reportlab~=4.4.1
httpx~=0.28.1